Put Arity outside of the edge type
